issue fixed

+++ database/migrations/2020_01_15_034438_create_revenues_table.php
@@ -26,7 +26,7 @@ class CreateRevenuesTable extends Migration
             $table->text('description')->nullable();
             $table->string('currency_code')->nullable()->default(null);
             $table->string('currency_symbol')->nullable()->default(null);
-            $table->float('exchange_rate')->default(0);
+            $table->double('exchange_rate')->default(0);
             $table->integer('created_by')->default('0');
             $table->unsignedBigInteger('created_user');
             $table->timestamps();
