+++ database/migrations/2020_04_02_045834_create_proposals_table.php
@@ -27,7 +27,7 @@ class CreateProposalsTable extends Migration
             $table->integer('converted_invoice_id')->default('0');
             $table->string('currency_code')->nullable()->default(null);
             $table->string('currency_symbol')->nullable()->default(null);
-            $table->float('exchange_rate')->default(0);
+            $table->double('exchange_rate')->default(0);
             $table->integer('created_by')->default('0');
             $table->unsignedBigInteger('created_user');
             $table->timestamps();
